requirements.txt
click==8.1.1
Flask==2.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
Werkzeug==2.1.0
